+++ examples/html/script.js
@@ -1,0 +1,7 @@
+window.addEventListener("contextmenu", (e) => {
+    e.preventDefault();
+});
+
+document.body.addEventListener("contextmenu", (e) => {
+    console.log("Own contextmenu");
+});