+++ examples/window_updated.py
@@ -13,7 +13,7 @@ def main() -> None:
     window = appverse.create_window(
         width=960,
         height=640,
-        debug=True,
+        debug=False,
         devtools=False,
     )